# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Innowise E-Commerce API",
        "description": "API documentation for Innowise E-Commerce Solutions",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get current cart contents",
                "description": "Returns cart items (identified by user ID for authenticated users, or session cookie for guests) along with subtotal, tax, and total.",
                "operationId": "7fdb9b40bd4ad2baf12aa15164eab788",
                "responses": {
                    "200": {
                        "description": "Cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "product": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "totals": {
                                            "properties": {
                                                "subtotal": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 199.98
                                                },
                                                "tax": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 199.98
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a product to the cart",
                "description": "Adds the given product to the cart. If the product already exists in the cart, the quantity is incremented.",
                "operationId": "cf9633893102fd28e9f7fa0e8bff7f36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item added to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear all items from the cart",
                "operationId": "2c37760f7c1c679d2a0c391e6a666105",
                "responses": {
                    "200": {
                        "description": "Cart cleared"
                    }
                }
            }
        },
        "/api/cart/{id}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Update the quantity of a cart item",
                "description": "Updates quantity. If quantity is 0, the item is removed from the cart.",
                "operationId": "a90eb30a1b5bc009fcdf69981c6ed0f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item updated or removed"
                    },
                    "404": {
                        "description": "Cart item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove a specific item from the cart",
                "operationId": "51e39afcdb34a29d0ec3feb88ff6b8d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cart item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed from cart"
                    },
                    "404": {
                        "description": "Cart item not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CartResource": {
                "description": "Cart item resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "product": {
                        "$ref": "#/components/schemas/ProductResource"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "description": "Product resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "category_name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "The latest Apple flagship phone."
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "old_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1099.99,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "url": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://cdn.example.com/products/iphone15.jpg"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}